# docs/openapi_admin.yaml
openapi: 3.0.0
info:
  title: Admin API Documentation
  description: Dokumentasi API khusus untuk Web Admin Panel (Laravel)
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local Development Server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []
paths:
...
  /api/admin/banner:
    post:
      summary: Create Banner (Upload)
      tags: [Master Data]
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title: { type: string }
                foto: { type: string, format: binary }
      responses:
        '200':
          description: Banner created
  /api/admin/banner/{id}:
    delete:
      summary: Deactivate Banner (Soft Delete)
      tags: [Master Data]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Banner deactivated